Fullyy functional security with permissions, fully functional country api

+++ src/main/java/service/impl/CountryServiceImpl.java
@@ -3,11 +3,15 @@ package service.impl;
 import dao.CountryDAO;
 import dto.CountryDTO;
 import model.Country;
+import model.User;
+import security.AuthUtils;
+import security.AuthenticatedUser;
 import service.CountryService;
 
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 import javax.ws.rs.BadRequestException;
+import javax.ws.rs.NotAuthorizedException;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
@@ -20,9 +24,13 @@ public class CountryServiceImpl implements CountryService {
 
 
     @Override
-    public CountryDTO add(CountryDTO object) {
+    public CountryDTO add(CountryDTO object, User authUser) {
         try {
-            countryDao.add(convertToEntity(object, Country.class));
+
+            if(!AuthUtils.checkIfAdmin(authUser))
+                throw new NotAuthorizedException("Permission Denied");
+
+            object.setId(countryDao.add(convertToEntity(object, Country.class)));
             return object;
         } catch (SQLException e) {
             throw new BadRequestException(e.getMessage());
@@ -30,8 +38,12 @@ public class CountryServiceImpl implements CountryService {
     }
 
     @Override
-    public CountryDTO removeById(int id) {
+    public CountryDTO removeById(int id, User authUser) {
         try {
+
+            if(!AuthUtils.checkIfAdmin(authUser))
+                throw new NotAuthorizedException("Permission Denied");
+
             countryDao.removeById(id);
             CountryDTO dto = new CountryDTO();
             dto.setId(id);
@@ -42,8 +54,12 @@ public class CountryServiceImpl implements CountryService {
     }
 
     @Override
-    public CountryDTO update(CountryDTO object) {
+    public CountryDTO update(CountryDTO object, User authUser) {
         try {
+
+            if(!AuthUtils.checkIfAdmin(authUser))
+                throw new NotAuthorizedException("Permission Denied");
+
             countryDao.update(convertToEntity(object, Country.class));
             return object;
         } catch (SQLException e) {
